Experimentally added resume function.

+++ convert.py
@@ -1,18 +1,14 @@
 import os
 import shutil
 import sys
-import threading
-from collections import deque
-from threading import Thread
-from time import sleep
 
 import torch
 import tqdm
 from torch.utils.data import Dataset
-from torchvision import transforms
 
 from dataloader import Dataloader
 from model import Net
+from utils import Writer, ConvertSampler
 
 torch.backends.cudnn.benchmark = True
 torch.backends.cudnn.fastest = True
@@ -23,37 +19,62 @@ def dummy_collate(x):
 
 
 def convert(args):
-    # TODO: Speedups:
-    #   Save images in threads (after moved to CPU)
-
     temp_folder = 'temp'
-    if os.path.isdir(temp_folder):
-        shutil.rmtree(temp_folder, ignore_errors=True)
 
-    os.makedirs(temp_folder)
-    os.makedirs(f'{temp_folder}\\output')
+    dest = f'{temp_folder}\\output'
+    resume_index = 1
 
-    if args.input_video is not None:
-        os.makedirs(f'{temp_folder}\\input')
-        code = os.system(f"ffmpeg -i {args.input_video} -vsync 0 {temp_folder}\\input\\%9d.png")
-        if code:
-            print('Failed to convert video to images.')
-            sys.exit(1)
+    if args.resume:
+        print('Resuming conversion...')
+        if os.path.exists(f'{temp_folder}\\output'):
+            if len(os.listdir(f'{temp_folder}\\output')) > 5:
+                resume_index = (len(os.listdir(f'{temp_folder}\\output')) - 1) // (args.sf + 1)
+            else:
+                resume_index = 1
         else:
-            input_path = f'{temp_folder}\\input'
+            raise Exception('Nothing to resume! ')
+    else:
+        if os.path.isdir(temp_folder):
+            if os.path.exists(f'{temp_folder}\\output') and len(os.listdir(f'{temp_folder}\\output')):
+                raise Exception('Folder is already in use! Did you intend to resume the progress? Use the --resume flag')
+
+            elif args.image_folder is None:  # Remove if we are about to process a new video
+                shutil.rmtree(temp_folder, ignore_errors=True)
+                os.makedirs(temp_folder)
+                os.makedirs(f'{temp_folder}\\input')
+
+        else:
+            os.makedirs(temp_folder)
+            os.makedirs(f'{temp_folder}\\input')
+
+        os.makedirs(f'{temp_folder}\\output')
+
+    # Setup conditions depending on video or image input
+    if args.input_video is not None:
+        if not args.resume:
+            code = os.system(f"ffmpeg -i {args.input_video} -vsync 0 {temp_folder}\\input\\%9d.png")
+
+            if code:
+                print('Failed to convert video to images.')
+                sys.exit(1)
+
+        input_path = f'{temp_folder}\\input'
 
     elif args.image_folder is not None:
         input_path = args.image_folder
     else:
         raise KeyError('Either input video or images folder required!')
 
+    # Create output folder
+
     use_cuda = not args.no_cuda and torch.cuda.is_available()
     if use_cuda:
         print('Cuda enabled!')
 
     dataset = Dataloader(path=input_path, cuda=use_cuda, train=False)
-    testloader = torch.utils.data.DataLoader(dataset, batch_size=1, shuffle=False, pin_memory=False,
-                                             collate_fn=dummy_collate)
+    testloader = torch.utils.data.DataLoader(dataset, sampler=ConvertSampler(dataset, resume_index-1),
+                                             batch_size=1, shuffle=False, pin_memory=False,
+                                             collate_fn=dummy_collate, num_workers=0)
     model = Net()
 
     for i in reversed(os.listdir('models')):
@@ -70,17 +91,18 @@ def convert(args):
     model.eval()
     intermediates = args.sf  # sf in superslomo
 
-    dest = f'{temp_folder}\\output'
     writer = Writer()
     writer.start()
-
+    print('folder len', len(os.listdir(f'{temp_folder}\\output')))
+    print('idx', resume_index)
     with torch.no_grad():
-        img_count = 1
+        img_count = resume_index + resume_index * args.sf - args.sf
+        print('img count', img_count)
+        print('_________')
+
         for data in tqdm.tqdm(testloader):
             img1, img2, img1_data, img2_data = data[0]
             del data
-            img1 = img1.unsqueeze(0)
-            img2 = img2.unsqueeze(0)
 
             if img_count == 1:
                 writer.add_job('copy',
@@ -105,8 +127,9 @@ def convert(args):
             img_count += intermediates + 1
             # print(f'\rProcessed {img_count} images', flush=True, end='')
     writer.exit_flag = True
-    while writer.is_alive():
-        sleep(0.1)
+    print('Waiting for file io to end...')
+    writer.join()
+    print('File io done!')
 
     code = os.system(
         f"ffmpeg -r {args.fps} -y -i {temp_folder}\\output\\%9d.png -an -c:v libvpx-vp9 -tile-columns 2 -tile-rows 1 -threads 12 -row-mt 1 -static-thresh 0 -frame-parallel 0 -auto-alt-ref 6 -lag-in-frames 25 -g 120 -crf 30 -pix_fmt yuv420p -cpu-used 4 -b:v 20M  -f webm -passlogfile ffmpeg2pass93057 -pass 1 NUL")
@@ -115,7 +138,7 @@ def convert(args):
         sys.exit(code)
 
     code = os.system(
-        f"ffmpeg -r {args.fps} -i {temp_folder}\\output\\%9d.png -c:a copy -c:v libvpx-vp9 -tile-columns 2 -tile-rows 1 -threads 12 -row-mt 1 -static-thresh 0 -frame-parallel 0 -auto-alt-ref 6 -lag-in-frames 25 -g 120 -crf 30 -pix_fmt yuv420p -cpu-used 1 -b:v 20M  -f webm -passlogfile ffmpeg2pass93057 -pass 2 {args.output_video}")
+        f"ffmpeg -r {args.fps} -y -i {temp_folder}\\output\\%9d.png -c:a copy -c:v libvpx-vp9 -tile-columns 2 -tile-rows 1 -threads 12 -row-mt 1 -static-thresh 0 -frame-parallel 0 -auto-alt-ref 6 -lag-in-frames 25 -g 120 -crf 30 -pix_fmt yuv420p -cpu-used 1 -b:v 20M  -f webm -passlogfile ffmpeg2pass93057 -pass 2 {args.output_video}")
 
     if not code:
         pass  # shutil.rmtree({temp_folder}, ignore_errors=True)
